# Remove requirement to bit library in colour module.