@1bardesign 2022-05-18
will need to find some time to write tests to ensure this is working as intended!

the amount of modulos make me a bit nervous performance wise though vs the bit operations, especially when compiled. i think i'd prefer a flag to use the bit library for speed rather than just removing the dependency entirely.