@1bardesign 2022-09-13
we can consider something (global or inline local config var) to force no bitops if required :+1: